[fix] fourth attempt fixing 404 error due to trailing slash
After three attempts of solving for the trailing slash in URL throwing 404,it is clear that the issue is with the fact that I had applied a permalink in each post's frontmatter. So using clever find and replace, removed the permalink attribute from all 186 posts available.

This is how done using Find and Replace:

Go to Find and Replace, enable Regex mode by clicking .* and add this: ^permalink: .+$